File changes for server migration & README edits
Due to problems faced with the current server, application is migrated
over to previous server, which is more stable. These file changes are to
account for the change in database logins as well as other fixes due to
the migration of application to different server.

README has been modified to reflect new application URL.